+++ lib/src/authsignal_push.dart
@@ -28,10 +28,20 @@ class AuthsignalPush {
     }
   }
 
-  Future<AuthsignalResponse<AppCredential>> addCredential({String? token}) async {
+  Future<AuthsignalResponse<AppCredential>> addCredential({
+    String? token,
+    bool requireUserAuthentication = false,
+    KeychainAccess? keychainAccess,
+    bool performAttestation = false,
+  }) async {
     await initCheck();
 
-    var arguments = <String, dynamic>{'token': token};
+    final arguments = <String, dynamic>{
+      'token': token,
+      'requireUserAuthentication': requireUserAuthentication,
+      'keychainAccess': keychainAccess?.value,
+      'performAttestation': performAttestation,
+    };
 
     try {
       final data = await methodChannel